+++ examples/src/main/kotlin/cl/ravenhill/keen/ga/knapsack/unbounded/Main.kt
@@ -1,0 +1,19 @@
+/*
+ * Copyright (c) 2023, Ignacio Slater M.
+ * 2-Clause BSD License.
+ */
+
+
+package cl.ravenhill.keen.ga.knapsack.unbounded
+
+import cl.ravenhill.keen.listeners.EvolutionPlotter
+import cl.ravenhill.keen.listeners.EvolutionSummary
+
+fun main() {
+    val summary = EvolutionSummary<Pair<Int, Int>, KnapsackGene>()
+    val plotter = EvolutionPlotter<Pair<Int, Int>, KnapsackGene>()
+    UnboundedKnapsackProblem(summary, plotter)
+    summary.display()
+    summary.fittest.genotype.flatten().filter { it.first != 0 }.also(::println)
+    plotter.display()
+}